requirements.txt
# No external libraries needed for core functionality

# Required dependencies for Streamlit app
streamlit==1.45.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.20.0
matplotlib>=3.5.0
networkx>=3.0.0
graphviz>=0.20.1
blinker==1.9.0
cachetools==5.5.2

# LLM integration dependencies
transformers>=4.30.0
torch>=2.0.0
python-dotenv>=1.0.0

# Visualization and export tools
pydot>=1.4.2
markdown>=3.4.1
PyPDF2>=3.0.0



# Optional dependencies for additional features
# Uncomment if needed:
# markdown>=3.4.1     # For markdown report generation